python solved

+++ 대회/2021 POSTECH Programming Open Contest/회전/app.c
@@ -3,14 +3,14 @@
 #include <string.h>
 #define BUF_SIZE 15
 
-int solve(int num_data, char *num_str);
-int add(int num, int len);
+int solve(unsigned long long int num_data, char *num_str);
+int add(unsigned long long int num, int len);
 
 int main(void)
 {
-    int num = 0;
+    unsigned long long int num = 0;
     int num_length = 0;
-    int total = 0;
+    unsigned long long int total = 0;
     scanf("%d", &num);
     char buf[BUF_SIZE];
     sprintf(buf, "%d", num);
@@ -26,16 +26,16 @@ int main(void)
     return 0;
 }
 
-int solve(int num_data, char *num_str)
+int solve(unsigned long long int num_data, char *num_str)
 {
     int num_length = strlen(num_str);
-    int num1 = add(num_data % 10, num_length);
-    int num2 = num_data / 10;
+    unsigned long long int num1 = add(num_data % 10, num_length);
+    unsigned long long int num2 = num_data / 10;
 
     return num1 + num2;
 }
 
-int add(int num, int len)
+int add(unsigned long long int num, int len)
 {
     for (int i = 0; i < len - 1; ++i)
     {
@@ -43,3 +43,14 @@ int add(int num, int len)
     }
     return num;
 }
+
+// #include <stdio.h>
+
+// int main(void)
+// {
+//     unsigned long long int num = 987654321;
+//     // int num = 9000000000;
+//     printf("%u", num);
+
+//     return 0;
+// }
